LayoutTests/editing/pasteboard/paste-delete-insertion-position-skip-paragraph-expected.txt

This test ensures CyberKit does not skip paragraphs when determining the insertion position for paste.
You should see "hello world<caret> CyberKit" below.

Initial markup:
| "
"
| <font>
|   class="Apple-style-span"
|   face="'courier new', monospace"
|   <div>
|     style="font-family: arial;"
|     "
"
|     <span>
|       class="Apple-style-span"
|       style="font-family: 'courier new', monospace;"
|       "hello"
|   "
"
|   <div>
|     style="font-family: arial; "
|     <span>
|       class="Apple-style-span"
|       style="font-family: 'courier new', monospace;"
|       " CyberKit"
| "
"

After inserting " world":
| "
"
| <font>
|   class="Apple-style-span"
|   face="'courier new', monospace"
|   <div>
|     style="font-family: arial;"
|     "
"
|     <span>
|       class="Apple-style-span"
|       style="font-family: 'courier new', monospace;"
|       "hello"
|   "
"
|   <div>
|     style="font-family: arial; "
|     <span>
|       class="Apple-style-span"
|       style="font-family: 'courier new', monospace;"
|       " CyberKit"
| "
"

After deleting line break before " CyberKit":
| "
"
| <font>
|   class="Apple-style-span"
|   face="'courier new', monospace"
|   <div>
|     style="font-family: arial;"
|     "
"
|     <span>
|       class="Apple-style-span"
|       style="font-family: 'courier new', monospace;"
|       "hello"
|   "
"
|   <div>
|     style="font-family: arial; "
|     <span>
|       class="Apple-style-span"
|       style="font-family: 'courier new', monospace;"
|       " CyberKit"
| "
"
